更新数据库应该不影响你们

+++ src/main/java/com/hm/sys/service/impl/SysAccountServiceImpl.java
@@ -186,7 +186,7 @@ public class SysAccountServiceImpl implements SysAccountService {
 		}
 
 		for (CheckInfo ci : checkList) {
-			Float checkinMoney = ci.getPaidUpMoney();
+			Double checkinMoney = ci.getPaidUpMoney();
 
 			if (checkinMoney != null)
 				checkinCount += checkinMoney;
